scripts/llm/requirements.txt
azure-identity
langchain
openai
pandas
python-dotenv
requests
tiktoken
torch
transformers>=4.34.0
langchain_openai
langchain-community